cryptfs-tpm2: fix GRUB2 long decrypt time for encrypted /boot
The grub will decode slots 0 and 1. It doesn't know in advance which
one will decrypt with the key file you provided so it will test them all
until it manages to decrypt a slot[1]. Place the newly added key file to
slot 0 to save search time for grub. And place original tpm key to slot 1.

Set --iter-time 10 for the newly added key file to save the decrypt
time for grub. Here is the manual[2][3][4]:

   --key-slot, -S <0-N>
       For LUKS operations that add key material, this option allows
       you to specify which key slot is selected for the new key.

   --new-key-slot <0-N>
       This option allows you to specify which key slot is selected
       for the new key.

   --iter-time, -i <number of milliseconds>
       The number of milliseconds to spend with PBKDF passphrase
       processing. Specifying 0 as parameter selects the compiled-in
       default.

[1] https://bbs.archlinux.org/viewtopic.php?id=228865
[2] https://man7.org/linux/man-pages/man8/cryptsetup-luksformat.8.html
[3] https://man7.org/linux/man-pages/man8/cryptsetup-luksOpen.8.html
[4] https://man7.org/linux/man-pages/man8/cryptsetup-luksAddKey.8.html

Issue: LIN1023-739
(LOCAL REV: NOT UPSTREAM) -- Will send upstream later

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>