cryptfs-tpm2: use PCR 23 for TMP2 to encrypt storage
Original PCR 7 is used for Secure Boot Policy [1][2]

The benefit of anchoring the TPM is that the machine status cannot be
compromised by any attack. If compromised, the system cannot boot up
due to the failure when mouting the rootfs, or access the encrypted partition
when mounting the LUKS partition. This is caused by the fact that the content
of PCR 7 is different with the value when creating the encrypted storage.
Usually, the content of PCR 7 is calculated based on the status of UEFI
secure boot. [3]

If use PCR 7 for TPM2 to encrypt disk while secure boot is disable, and
then enable secure boot, the PCR7 is forbidden to use for TPM2 due to
status of UEFI secure boot is changed. Use PCR 23 for TPM2 as workaround
for this situation

PCR 23 is application support that can be set and reset by the OS [1]

Add a distro feature tpm-pcr-23 to apply the patch conditionally

[1] https://wiki.gentoo.org/wiki/Trusted_Platform_Module
[2] https://github.com/jiazhang0/cryptfs-tpm2/blob/master/README#L72
[3] https://github.com/jiazhang0/meta-secure-core/tree/master/meta-encrypted-storage#mount-the-luks-partition

Issue: LIN1023-1293
(LOCAL REV: NOT UPSTREAM) -- Will send upstream later

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>