tpm2-tools: fix CVE-2024-29038
Issue:LIN1023-5023

A flaw was found in the tpm2-tools package. This issue occurs due to a
missing check whether the magic number in attest is equal to TPM2_GENERATED_VALUE,
which can allow an attacker to generate arbitrary quote data that may not be
detected by tpm2_checkquote.

References:
https://ubuntu.com/security/CVE-2024-29038

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance
Signed-off-by: Meenali Gupta <meenali.gupta@windriver.com>